Redid page system to work correctly.
Before the pages were determined by searching ID, however this lead to a problem where if there was a missing ID it would lead to empty pages. (example: ID 1 to 50 exist, 51 to 101 don't exist, ID 102 to w/e exist, page 1 would be full, page 2 would be empty, and page 3 would be full. This has been corrected.